+++ Orleans.Redis.sln
@@ -3,15 +3,15 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2036
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Storage.Redis", "Orleans.Storage.Redis\Orleans.Storage.Redis.csproj", "{C64DA14D-16F0-4DC9-B329-FD71F6F2C8CB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Persistence.Redis", "Orleans.Persistence.Redis\Orleans.Persistence.Redis.csproj", "{C64DA14D-16F0-4DC9-B329-FD71F6F2C8CB}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{AD3FE114-8745-460B-B200-7F0A3F3CA53F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Storage.Redis.TestGrains", "Tests\Orleans.Storage.Redis.TestGrains\Orleans.Storage.Redis.TestGrains.csproj", "{F401E365-B435-4970-8A54-F8D55161DCC0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Persistence.Redis.TestGrains", "Tests\Orleans.Persistence.Redis.TestGrains\Orleans.Persistence.Redis.TestGrains.csproj", "{F401E365-B435-4970-8A54-F8D55161DCC0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Storage.Redis.TestGrainInterfaces", "Tests\Orleans.Storage.Redis.TestGrainInterfaces\Orleans.Storage.Redis.TestGrainInterfaces.csproj", "{163837CC-AE6E-4BCD-A0F1-B67659E45695}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Persistence.Redis.TestGrainInterfaces", "Tests\Orleans.Persistence.Redis.TestGrainInterfaces\Orleans.Persistence.Redis.TestGrainInterfaces.csproj", "{163837CC-AE6E-4BCD-A0F1-B67659E45695}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Storage.Redis.Tests", "Tests\Orleans.Storage.Redis.Tests\Orleans.Storage.Redis.Tests.csproj", "{055EA4B7-201B-4451-8368-1B41B26C28A8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Persistence.Redis.Tests", "Tests\Orleans.Persistence.Redis.Tests\Orleans.Persistence.Redis.Tests.csproj", "{055EA4B7-201B-4451-8368-1B41B26C28A8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
