+++ Tests/Orleans.Persistence.Redis.TestGrains/BinaryTestGrain.cs
@@ -1,10 +1,10 @@
 ﻿using Orleans.Providers;
 using Orleans.Runtime;
-using Orleans.Storage.Redis.TestGrainInterfaces;
+using Orleans.Persistence.Redis.TestGrainInterfaces;
 using System;
 using System.Threading.Tasks;
 
-namespace Orleans.Storage.Redis.TestGrains
+namespace Orleans.Persistence.Redis.TestGrains
 {
     [StorageProvider(ProviderName = "REDIS-BINARY")]
     public class BinaryTestGrain : Grain<BinaryTestGrainState>, IBinaryTestGrain
