+++ Tests/Orleans.Persistence.Redis.TestGrains/BinaryTestGrainState2.cs
@@ -1,9 +1,7 @@
-﻿using Orleans.Storage.Redis.TestGrainInterfaces;
+﻿using Orleans.Persistence.Redis.TestGrainInterfaces;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Orleans.Storage.Redis.TestGrains
+namespace Orleans.Persistence.Redis.TestGrains
 {
     public class BinaryTestGrainState2
     {
