+++ Tests/Orleans.Persistence.Redis.TestGrains/JsonTestGrain2.cs
@@ -1,10 +1,10 @@
 ﻿using Orleans.Providers;
 using Orleans.Runtime;
-using Orleans.Storage.Redis.TestGrainInterfaces;
+using Orleans.Persistence.Redis.TestGrainInterfaces;
 using System;
 using System.Threading.Tasks;
 
-namespace Orleans.Storage.Redis.TestGrains
+namespace Orleans.Persistence.Redis.TestGrains
 {
     [StorageProvider(ProviderName = "REDIS-JSON")]
     public class JsonTestGrain2 : Grain<JsonTestGrainState2>, IJsonTestGrain2
