+++ Tests/Orleans.Persistence.Redis.TestGrains/JsonTestGrainState2.cs
@@ -1,11 +1,9 @@
-﻿using Orleans.Storage.Redis.TestGrainInterfaces;
+﻿using Orleans.Persistence.Redis.TestGrainInterfaces;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Orleans.Storage.Redis.TestGrains
+namespace Orleans.Persistence.Redis.TestGrains
 {
-    public class JsonTestGrainState
+    public class JsonTestGrainState2
     {
         public string StringValue { get; set; }
         public int IntValue { get; set; }
